+++ product/dashboard/src/webapp/config/webpack.dev.js
@@ -2,11 +2,12 @@ process.env.CODEKVAST_VERSION = 'dev';
 process.env.ENV = 'development';
 
 var webpackMerge = require('webpack-merge');
-var ExtractTextPlugin = require('extract-text-webpack-plugin');
 var commonConfig = require('./webpack.common.js');
 var helpers = require('./helpers');
 
 module.exports = webpackMerge(commonConfig, {
+    mode: 'development',
+
     devtool: 'cheap-module-eval-source-map',
 
     output: {
@@ -15,10 +16,6 @@ module.exports = webpackMerge(commonConfig, {
         filename: '[name].js',
         chunkFilename: '[id].chunk.js'
     },
-
-    plugins: [
-        new ExtractTextPlugin('[name].css')
-    ],
 
     devServer: {
         inline: true,